Move upstream remote to justfile
Refactors the fetch command to instead compare two commit hashes, rather than comparing the "upstream" branch and the remote head branch. The latter required some finagling because you cannot necessarily be sure that it'll ALWAYS be "origin/master", so it was delegated to each submodule's reset script. However, since we know that the remote will ALWAYS be origin, it seems that "refs/remotes/origin/HEAD" is a reliable way to do the same thing.